Implement pagination for listing page

+++ src/main/java/com/thedevs/real_estate/dto/CreateInquiryRequest.java
@@ -2,14 +2,14 @@ package com.thedevs.real_estate.dto;
 
 public class CreateInquiryRequest {
 
-    private Long listingId;
+    private Integer listingId;
     private String message;
 
-    public Long getListingId() {
+    public Integer getListingId() {
         return listingId;
     }
 
-    public void setListingId(Long listingId) {
+    public void setListingId(Integer listingId) {
         this.listingId = listingId;
     }
